0.0.8/maintenance: yard adjustment and some refactoring
