+++ lib/sf_cli/sf/apex/run.rb
@@ -4,19 +4,25 @@ require 'stringio'
 module SfCli::Sf::Apex
   module Run
     #
-    # run apex code and returns its result
+    # Run apex code and returns its result.
+    # If you don't specify the script file, it starts interactive mode.
+    # @see https://developer.salesforce.com/docs/atlas.en-us.sfdx_cli_reference.meta/sfdx_cli_reference/cli_reference_apex_commands_unified.htm#cli_reference_apex_run_unified CLI reference
     #
-    # *target_org* --- an alias of paticular org, or username can be used<br>
+    # @param target_org [Symbol,String]
+    #   an alias of paticular org, or username can be used.
     #
-    # *file* --- (1) path to a local file that contains \Apex code. (2) StringIO object
+    # @param file [String,StringIO]
+    #   (1) path to a local file that contains Apex code. 
+    #   (2) StringIO object
     #
-    # == Examples
-    # Execute apex code in a file
+    # @return [ApexResult] Apex execution result.
+    #
+    # @example Execute apex code in a file
     #   result = sf.apex.run file: "path/to/apex"
     #   result.success # true if the execution succeeds
     #   result.logs    # execution log
     #
-    # StringIO is usable instead of file path
+    # @example StringIO is usable instead of file path
     #   require 'stringio'
     #
     #   pseudo_file = StringIO.new <<~EOS
@@ -26,9 +32,7 @@ module SfCli::Sf::Apex
     #
     #   sf.apex.run target_org: :dev, file: pseudo_file
     #
-    # == Interactive Mode
-    #
-    # If you don't specify *file* argument, it starts interactive mode that may be helpful in IRB environment.
+    # @example Interactive Mode
     #   irb(main:) > sf.apex.run target_org: :dev
     #
     #   Account acc = [SELECT Id, Name FROM Account LIMIT 1];
@@ -48,7 +52,6 @@ module SfCli::Sf::Apex
     #       "Execute Anonymous: Account acc = [SELECT Id, Name FROM Account LIMIT 1];",
     #       "Execute Anonymous: System.debug(acc.Name);",
     #      ....]
-    # For more command details, see {the command reference}[https://developer.salesforce.com/docs/atlas.en-us.sfdx_cli_reference.meta/sfdx_cli_reference/cli_reference_apex_commands_unified.htm#cli_reference_apex_run_unified]
     #
     def run(target_org: nil, file: nil)
       return run_interactive(target_org) if file.nil?
