+++ lib/sf_cli/sf/data/create_record.rb
@@ -1,18 +1,16 @@
 module SfCli::Sf::Data
   module CreateRecord
     # create a object record.
+    # @see https://developer.salesforce.com/docs/atlas.en-us.sfdx_cli_reference.meta/sfdx_cli_reference/cli_reference_data_commands_unified.htm#cli_reference_data_create_record_unified command reference
+    # @param object_type [Symbol, String] object type(ex. Account)
+    # @param values      [Hash]           field values to be assigned
+    # @param target_org  [Symbol, String] an alias of paticular org, or username can be used
     #
-    # *object_type* --- \Object Type (ex. Account)<br>
+    # @return [String] record ID
     #
-    # *values* --- field values to be assigned<br>
-    #
-    # *target_org* --- an alias of paticular org, or username can be used<br>
-    #
-    # ======
+    # @example
     #   # create a TheCustomObject record with name and age
     #   sf.data.create_record :TheCustomObject__c, values: {Name: "John Smith", Age: 33}
-    #
-    # For more command details, see {the command reference}[https://developer.salesforce.com/docs/atlas.en-us.sfdx_cli_reference.meta/sfdx_cli_reference/cli_reference_data_commands_unified.htm#cli_reference_data_create_record_unified]
     #
     def create_record(object_type, values: {}, target_org: nil)
       field_values = field_value_pairs(values)
