+++ lib/sf_cli/sf/data/helper_methods.rb
@@ -1,6 +1,7 @@
 module SfCli
   module Sf
     module Data
+      # @private
       module HelperMethods
         def prepare_record(hash) # :doc:
           hash.delete 'attributes'
