+++ lib/sf_cli/sf/data/upsert_bulk.rb
@@ -2,19 +2,16 @@ require_relative './bulk_result_v2'
 
 module SfCli::Sf::Data
   module UpsertBulk
-    # update records using Bulk API 2.0
+    # Update records using Bulk API 2.0
+    # @param file        [String]        path of a CSV file to update records
+    # @param sobject     [Symbol,String] object type(ex. Account)
+    # @param external_id [String]        name of the external ID field.Otherwise it must be Id
+    # @param timeout     [Integer]       max minutes to wait for the job complete the task
+    # @param target_org [Symbol,String]  an alias of paticular org, or username can be used
     #
-    # *file* --- a \CSV file for update records<br>
+    # @return [JobInfo, BulkResultV2] the job result, whose type is changed by situation
     #
-    # *sobject* --- \Object Type (ex. Account)<br>
-    #
-    # *external_id* --- Name of the external ID field, or the Id field<br>
-    #
-    # *timeout* --- max minutes to wait for the job complete the task.<br>
-    #
-    # *target_org* --- an alias of paticular org, or username can be used<br>
-    #
-    # ======
+    # @example
     #   # start a upsert job
     #   jobinfo = sf.data.upsert_bulk sobject: :TestCustomObject__c, file: 'upsert.csv' # this returns immediately
     #   jobinfo.id  # => "750J4000003g1OaIAI" it's job ID
@@ -25,7 +22,7 @@ module SfCli::Sf::Data
     #   # Or, you can wait for the job completion with one try.
     #   result = sf.data.upsert_bulk sobject: :TestCustomObject__c, file: 'upsert.csv', timeout: 5  # wait within 5 minutes
     #
-    # For more command details, see {the command reference}[https://developer.salesforce.com/docs/atlas.en-us.sfdx_cli_reference.meta/sfdx_cli_reference/cli_reference_data_commands_unified.htm#cli_reference_data_upsert_bulk_unified]
+    # @see https://developer.salesforce.com/docs/atlas.en-us.sfdx_cli_reference.meta/sfdx_cli_reference/cli_reference_data_commands_unified.htm#cli_reference_data_upsert_bulk_unified command reference
     #
     def upsert_bulk(file:, sobject:, external_id:, timeout: nil, target_org: nil)
       flags = {
