+++ lib/sf_cli/sf/data/upsert_resume.rb
@@ -2,15 +2,14 @@ require_relative './bulk_result_v2'
 
 module SfCli::Sf::Data
   module UpsertResume
-    # resume a bulk upsert job you previously started with Bulk API 2.0 and return a bulk result object.
+    # Resume a bulk upsert job you previously started with Bulk API 2.0 and return a bulk result object.
+    # @param job_id     [String]        job ID you want to resume
+    # @param timeout    [Integer]       max minutes to wait for the job complete the task
+    # @param target_org [Symbol,String] an alias of paticular org, or username can be used
     #
-    # *job_id* ---  job ID you want to resume<br>
+    # @return [JobInfo, BulkResultV2] the job result, whose type is changed by situation
     #
-    # *timeout* --- max minutes to wait for the job complete the task.<br>
-    #
-    # *target_org* --- an alias of paticular org, or username can be used<br>
-    #
-    # ======
+    # @example
     #   # start a upsert job
     #   jobinfo = sf.data.upsert_bulk sobject: :TestCustomObject__c, file: 'upsert.csv' # this returns immediately
     #   jobinfo.id  # => "750J4000003g1OaIAI" it's job ID
@@ -22,9 +21,7 @@ module SfCli::Sf::Data
     #
     #   puts 'yey!' if result.job_info.completed? # the job has completed
     #
-    # To know more about a job result, take a look at SfCli::Sf::Data module
-    #
-    # For more command details, see {the command reference}[https://developer.salesforce.com/docs/atlas.en-us.sfdx_cli_reference.meta/sfdx_cli_reference/cli_reference_data_commands_unified.htm#cli_reference_data_upsert_resume_unified]
+    # @see https://developer.salesforce.com/docs/atlas.en-us.sfdx_cli_reference.meta/sfdx_cli_reference/cli_reference_data_commands_unified.htm#cli_reference_data_upsert_resume_unified command reference
     #
     def upsert_resume(job_id:, timeout: nil, target_org: nil)
       flags = {
