+++ lib/sf_cli/sf/main.rb
@@ -3,14 +3,15 @@ require 'json'
 
 module SfCli
   module Sf
-    # ==== description
-    # The entry class of sf command. It is returned by sf method.
+
+    # The entry class of sf command. It is returned by the sf method.
     # With including Singleton module, the instance of this class is guaranteed as singleton.
     #
-    # ==== examples
+    # @example
     #   sf             # returns a instance of SfCli::Sf
     #   sf.org.display # you can follow the similar syntax to the original command by using method chain.
     #
+    # @private :nodoc: just for developers
     class Main
       include Singleton
 
@@ -21,6 +22,9 @@ module SfCli
         attr_reader category.downcase.to_sym
       end
 
+      # Generate each sub command object such as org, data and sobject
+      # so that it can chain like `sf.org.display`.
+      #
       def initialize
         OPERATION_CATEGORIES.each do |category|
           instance_variable_set(:"@#{category.downcase}", Object.const_get(%|::SfCli::Sf::#{category}::Core|).new)
